feat: add server.ts
