+++ components/package.json
@@ -4,11 +4,10 @@
   "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "NODE_ENV=development node server",
+    "start": "ts-node --esm src/server.ts",
     "build:client": "vite build --outDir dist/client",
     "build:server": "vite build --ssr src/entry-server.tsx --outDir dist/server",
-    "build": "npm run build:client && npm run build:server",
-    "start": "node src/server.js"
+    "build": "npm run build:client && npm run build:server && tsc src/server.ts"
   },
   "dependencies": {
     "@reduxjs/toolkit": "^1.9.3",
@@ -20,7 +19,9 @@
   "devDependencies": {
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
+    "@types/express": "^4.17.17",
     "@types/jest": "^29.5.0",
+    "@types/node": "^18.16.3",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "@typescript-eslint/eslint-plugin": "^5.22.0",
@@ -44,6 +45,7 @@
     "react-router-dom": "^6.9.0",
     "sass": "^1.59.3",
     "ts-jest": "^29.0.5",
+    "ts-node": "^10.9.1",
     "typescript": "^4.9.3",
     "vite": "^4.1.0",
     "vite-node": "^0.30.1",
